+++ app/src/test/java/app/knotwork/android/presentation/ui/chat/home/ChatHomeConsoleStreamingTest.kt
@@ -162,6 +162,7 @@ class ChatHomeConsoleStreamingTest {
         loadModelUseCase,
         mockk(relaxed = true),
         mockk(relaxed = true),
+        mockk(relaxed = true),
         pipelineRunRepository,
         runTraceRepository,
         mockk(relaxed = true),